Move the social sign on module to lib/
